order-service/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-dotenv
motor